add some missing `<code>` tags to index.html
Both `<code><a ...>...</a></code>` and `<a ...><code>...</code></a>` existed before this PR. Both still exist, and so far as I can tell, both render as intended.

Still, it might be better to settle on always `<code><a ...>...</a></code>` or always `<a ...><code>...</code></a>`. If so, I lean toward the latter.